desarrollo-render/myclase/requirements.txt
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==5.2.5
django-allauth==65.11.1
django-environ==0.12.0
djangorestframework
dnspython==2.7.0
exceptiongroup==1.3.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.31.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==20.1.0
idna==3.10
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
numpy==1.26.4
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
polars==1.32.3
psutil==7.0.0
psycopg2-binary==2.9.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
PyMsgBox==1.0.9
pynput==1.8.1
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-decouple
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sounddevice==0.5.2
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.187
ultralytics-thop==2.0.16
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.2.0
